+++ packages/react-form-state/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Fixed
+
+- `submit` now checks for the existence of `preventDefault` on the event passed in before calling it.
+
 ## [0.6]
 
 ### Added